% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddInfo2res.R
\name{AddInfo2res}
\alias{AddInfo2res}
\title{AddInfo2res}
\usage{
AddInfo2res(res, CNV2HighPvalue, CNV2LowPvalue, Class, BAlleleFreq, MyBAF,
  LogRRatio, SumPeaks, SDChr, MeanChr)
}
\value{
return data in data frame
}
\description{
AddInfo2res: read data.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

