% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Affy_convert.R
\name{affyConvert}
\alias{affyConvert}
\title{Affymetrix Convert}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
affyConvert(PathToCel, Outdir = "affyConvert", PathToPennCNV, AffyAssay)
}
\arguments{
\item{PathToCel:}{The path to per sample Affymetrix CEL files.}

\item{Outdir:}{Output directory for the processed data and conversion files.}

\item{AffyAssay:}{Character string defining the affymetix assay to be converted (Affy_6.0 or Affy_5.0).}
}
\value{
Data frame with predicted CNVs.
}
\description{
Affymetrix Convert: Calculate LRR and BAF from Affymetix CEL files
to allow for CNV detection with iPsych CNV.
}
\details{
Compatable with conversion from Affymetrix Assay_6.0 and
Assay_5.0. This function applies the instructions given for
conversion of Affymetrix data to allow for PennCNV detection from
http://penncnv.openbioinformatics.org/en/latest/user-guide/affy/
}
\examples{
# Download Affy6.0 example data from hapmap and run analysis
system("mkdir -p affy6.0_example")
system("wget ftp://anonymous@ftp.ncbi.nlm.nih.gov/hapmap/raw_data/hapmap3_affy6.0/CHEAP.tgz")
system("tar xvzf CHEAP.tgz -C affy6.0_example")
affyConvert(PathToCel = "affy6.0_example", Outdir = "affyConvert", PathToPennCNV = "/home/mydir/penncnv/", AffyAssay = "Affy_6.0")

# Download Affy5.0 example data from ... and run analysis
system("mkdir -p affy5.0_example")
system("wget http://www.affymetrix.com/support/downloads/data/GW_SNP5_GCOS_CEL_1.zip")
system("unzip GW_SNP5_GCOS_CEL_1.zip -d affy5.0_example")
affyConvert(PathToCel = "affy5.0_example", Outdir = "affyConvert", PathToPennCNV = "/home/mydir/penncnv/", AffyAssay = "Affy_5.0")
}
\author{
Johan Hilge Thygesen
}

