% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindMultipleFiles.R
\name{BindMultipleFiles}
\alias{BindMultipleFiles}
\title{BindMultipleFiles}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
BindMultipleFiles(PathRawData = ".", Pattern = "CNV.tab",
  recursive = TRUE)
}
\arguments{
\item{PathRawData:}{Full path where the files are located.}

\item{Pattern:}{The patter of file name. For example, *.tab}

\item{recursive:}{If should search for sub-folders.}
}
\value{
Return data in a data frame.
}
\description{
BindMultipleObjects: Read files and combine.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
All.CNV.Files <- BindMultipleFiles(PathRawData=".", Pattern="CNV.tab", recursive=TRUE)
}
\author{
Marcelo Bertalan.
}

