% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CheckAbnChr.R
\name{CheckAbnChr}
\alias{CheckAbnChr}
\title{CheckAbnChr}
\usage{

  CheckAbnChr(Path2RawFiles = "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/Version2",
  Cores = 40, Pattern = "*.txt$", skip = 10, NumFiles = "All")
}
\value{
Abnormal chromosomes
}
\description{
CheckAbnChr
}
\author{
Marcelo Bertalan
}

