% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningPeaks.R
\name{CleaningPeaks}
\alias{CleaningPeaks}
\title{CleaningPeaks}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
CleaningPeaks(tp)
}
\arguments{
\item{Tp:}{Unknown.}
}
\value{
Peaks for BAF.
}
\description{
CleaningPeaks: Remove noise from peaks in B Allele Frequency (BAF).
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.
}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

