% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertPenntoRCNVs.R
\name{ConvertPenntoRCNVs}
\alias{ConvertPenntoRCNVs}
\title{ConvertPenntoRCNVs}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
ConvertPenntoRCNVs(PennCNVFile = "Test_CNVs.txt")
}
\arguments{
\item{PennCNVFile:}{File path with CNVs called in PennCNV.}
}
\value{
Return output from PennCNV in data frame and a R-readable file with the extension ".Rtxt".
}
\description{
ConvertPenntoRCNVs: Converts PennCNV-output to R-readable file with the extension ".Rtxt" and reads in as dataframe.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown

}
\author{
Ida Sønderby, Marcelo Bertalan, Louise K. Hoeffding.
}

