% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineCNVType.R
\name{DefineCNVType}
\alias{DefineCNVType}
\title{DefineCNVType}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
DefineCNVType(tmp2)
}
\arguments{
\item{tmp2:}{Unknown.}
}
\value{
CNV type.
}
\description{
DefineCNVType: Determine true CNVs calls.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown

}
\author{
Marcelo Bertalan, Louise K. Hoeffding
}

