% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineStartAndStop.R
\name{DefineStartAndStop}
\alias{DefineStartAndStop}
\title{DefineStartAndStop}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
DefineStartAndStop(indx, subCNV, CHR, ID, CPT.Res)
}
\arguments{
\item{Indx:}{Unknown.}

\item{SubCNV:}{Unknown.}

\item{CHR:}{Character, select a specific chromosome to be analyzed.}

\item{ID:}{Unknown.}

\item{CPT.Res:}{Unknown.}
}
\value{
Data frame with start and stop positions for each CNV.
}
\description{
DefineStartAndStop: Define start and stop positions for a given CNV.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding
}

