% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefiningLogRRatio.R
\name{DefiningLogRRatio}
\alias{DefiningLogRRatio}
\title{DefiningLogRRatio}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
DefiningLogRRatio(res2)
}
\arguments{
\item{res2:}{Unknown.}
}
\value{
Classification for LRR.
}
\description{
DefiningLogRRatio: Define the Log R Ratio (LRR).
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

