% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateMockResults.R
\name{EvaluateMockResults}
\alias{EvaluateMockResults}
\title{EvaluateMockResults}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
EvaluateMockResults(MockCNVs, df, Cores = 1)
}
\arguments{
\item{MockCNV:}{Unknown.}

\item{Df:}{Data frame with predicted CNVs for each sample? Unknown.}

\item{Cores:}{Number of cores used, default = 1.}
}
\value{
Data frame with predicted CNVs.
}
\description{
EvaluateMockResults: Unknown.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

