% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateMyBAF.R
\name{EvaluateMyBAF}
\alias{EvaluateMyBAF}
\title{EvaluateMyBAF}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
EvaluateMyBAF(res, res2)
}
\arguments{
\item{res:}{Unknown.}

\item{res2.}{Unknown.}
}
\value{
Classification for BAF.
}
\description{
EvaluateMyBAF: Evaluate B Allele Frequency (BAF).
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

