% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCNVs.V4.R
\name{FilterCNVs.V4}
\alias{FilterCNVs.V4}
\title{iPsychCNVs}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
FilterCNVs.V4(CNVs = CNVs, MinNumSNPs = 20, Sample, ID = "Test",
  verbose = FALSE)
}
\value{
Data frame with CNVs and classification.
}
\description{
Filter CNVs: function to filter predicted CNVs and avoid high number of false positives.
}
\details{
The function receives a data frame with CNV information, ex: Chr, Start position, Stop position and Sample ID. Returns a data frame with CNVs and a Type difining if is bad or good.
}
\author{
Marcelo Bertalan
}

