% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterFromCNVs.R
\name{FilterFromCNVs}
\alias{FilterFromCNVs}
\title{FilterFromCNVs}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
FilterFromCNVs(CNVs, PathRawData, MinNumSNPs = 10, Source = "iPsychCNV",
  Skip = 0, Cores = 1)
}
\arguments{
\item{CNVs:}{Data frame with CNVs.}

\item{PathRawData:}{The path to the raw data files containing Log R Ratio (LRR) and B Allele Frequency (BAF) values. Example: "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/Version2".}

\item{MinNumSNPs:}{Minimum number of SNPs per CNV, default = 10.}

\item{Source:}{Which method is the original call, default = iPsychCNV.}

\item{Skip:}{Integer, the number of rows of the data file that should be skipped before beginning tp read the data. Use if file has comments, default = 0.}

\item{Cores:}{Number of cores to run in parallel, default = 1.}
}
\value{
Data frame with the estimated copy number for each chromosome.
}
\description{
FilterFromCNVs: Filter CNV from other methods.
}
\details{
Specifically designed to reduce false positive CNVs and handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
LongRoi <- MakeLongMockSample(Mean=c(-0.6, -0.3, 0.3, 0.6), Size=c(200, 400, 600))
# GADA
Sample <- read.table("LongMockSample.tab", sep="\\t", header=TRUE, stringsAsFactors=F)
Gada <- RunGada(Sample)
Gada.filter <- FilterFromCNVs(CNVs=Gada, PathRawData=".", MinNumSNPs=10, Source="Gada", Skip=0, Cores=1)
See iPsychCNV tutorial for more examples
http://biopsych.dk/iPsychCNV/tutorial.html
}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

