% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FindCNV.V4.R
\name{FindCNV.V4}
\alias{FindCNV.V4}
\title{FindCNV}
\usage{
FindCNV.V4(ID, MinNumSNPs, CNV)
}
\value{
A data frame with CNVs, with start and stop position, chr and etc. for each sample.
}
\description{
FindCNV: Find CNVs using Log.R.Ratio data by cpt.meanvar method.
}
\details{
Input: sample path, CNV in data frame format and min number of SNPs
}
\author{
Marcelo Bertalan
}

