% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FixCNVPosition_V3.R
\name{FixCNVPosition}
\alias{FixCNVPosition}
\title{FixCNVPosition}
\usage{
FixCNVPosition(Df, subCNV, MinNumSNPs, ID)
}
\value{
Data frame with CNV information.
}
\description{
FixCNVPosition: Trim the CNV position by distance of SNPs.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

