% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_si.R
\name{format_si}
\alias{format_si}
\title{format_si}
\source{
\url{http://www.moeding.net/archives/32-Metric-prefixes-for-ggplot2-scales.html}
}
\usage{
format_si(...)
}
\arguments{
\item{\dots}{additional arguments to format()}
}
\value{
A function to format a vector of strings using SI prefix notation
}
\description{
format_si: Format a vector of numeric values according to the International System of Units.
}
\details{
Format a vector of numeric values according to the International System of Units.
\code{\link{http://en.wikipedia.org/wiki/SI_prefix}}

Based on code by Ben Tupper (\code{\link{https://stat.ethz.ch/pipermail/r-help/2012-January/299804.html}})
}
\examples{
ggplot() + ... + scale_y_continuous(labels=format_si())

}
\author{
Stefan Möding
}

