% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCNVClass.R
\name{GetCNVClass}
\alias{GetCNVClass}
\title{GetCNVClass}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
GetCNVClass(Variable)
}
\arguments{
\item{Variable:}{Numeric, Unknown.}
}
\value{
return data in data frame.
}
\description{
GetCNVClass: Read data.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

