% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCNVFreq.R
\name{GetCNVFreq}
\alias{GetCNVFreq}
\title{GetCNVFreq}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
GetCNVFreq(df, Pheno, Cut, roi)
}
\arguments{
\item{df}{Data frame with CNVs, must have status1 as phenotype variable and ID as sample ID variable.}

\item{Pheno}{Data frame with phenotypes must have status1 as phenotype variable and ID as sample ID variable.}

\item{Cut}{Cutoff value for CNV mean. Only negative for now. Default Cuf=-0.3.}

\item{roi}{Regions of interest, object with information where should be the CNV you are looking for.}
}
\value{
return data in a data frame.
}
\description{
GetCNVFreq: Get CNV freq.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan.
}

