% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataVariables.R
\name{GetDataVariables}
\alias{GetDataVariables}
\title{GetDataVariables}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
GetDataVariables(Data)
}
\arguments{
\item{Data:}{Unknown.}
}
\value{
Hotspots - Unknown.
}
\description{
GetDataVariables: Unknown.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

