% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGenotypeInfo.R
\name{GetGenotypeInfo}
\alias{GetGenotypeInfo}
\title{GetGenotypeInfo}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
GetGenotypeInfo(tmpRaw)
}
\arguments{
\item{tmpRaw:}{Unknown.}
}
\value{
Data frame with predicted CNVs.
}
\description{
GetGenotypeInfo: Unknown.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

