% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIndxPositionFromChips.R
\name{GetIndxPositionFromChips}
\alias{GetIndxPositionFromChips}
\title{GetIndxPositionFromChips}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
GetIndxPositionFromChips(CNVs, Sample)
}
\arguments{
\item{CNVs:}{Data frame with information on CNVs. Unknown.}

\item{Sample:}{Unknown.}
}
\value{
Data frame with predicted CNVs.
}
\description{
GetIndxPositionFromChips: Unknown.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

