% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLocus.R
\name{GetLocus}
\alias{GetLocus}
\title{GetLocus}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
GetLocus(df)
}
\arguments{
\item{Df:}{Data frame with predicted CNVs for each sample.}
}
\value{
data frame with CNV locus.
}
\description{
GetLocus: Unknown.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

