% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMockValues.R
\name{GetMockValues}
\alias{GetMockValues}
\title{iGetMockValues}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
GetMockValues(Type = "Blood")
}
\arguments{
\item{Type:}{Unkown, default = Blood.}
}
\value{
Data frame with predicted CNVs.
}
\description{
GetMockValues: Unknown.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

