% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HotspotsCNV.R
\name{HotspotsCNV}
\alias{HotspotsCNV}
\title{CompressCNVsHighFreq}
\usage{
HotspotsCNV(df, Freq = 1, OverlapCutoff = 0.7, Cores = 1)
}
\value{
Hotspots
}
\description{
HotspotsCNV
}
\author{
Marcelo Bertalan
}

