% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HotspotsPrediction.R
\name{HotspotsPrediction}
\alias{HotspotsPrediction}
\title{HotspotsPrediction}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
HotspotsPrediction(hotspots, df)
}
\arguments{
\item{Hotspots:}{Unkown.}

\item{df:}{Data frame. Unknown.}
}
\value{
Data frame with predicted hotspots.
}
\description{
HotspotsPrediction: Count number of true CNVs in each hotspots.
}
\details{
Unknown
}
\examples{
Unknown. 

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

