% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeAndEvalMockData.R
\name{MakeAndEvalMockData}
\alias{MakeAndEvalMockData}
\title{MakeAndEvalMockData}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
MakeAndEvalMockData(N = 100, Wave1 = FALSE, Type = "PKU", Cores = 20)
}
\arguments{
\item{N:}{Unknown, default = 100.}

\item{Wave1:}{Unknown, default = FALSE.}

\item{Type:}{Unknown, default = PKU.}

\item{Cores:}{Number of cores used, default = 20.}
}
\value{
Data frame with predicted CNVs.
}
\description{
MakeAndEvalMockData: Generate and evaulate Mock data.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown. 

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

