% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MakeOneMockSample.R
\name{MakeOneMockSample}
\alias{MakeOneMockSample}
\title{MakeOneMockSample}
\usage{
MakeOneMockSample(Noise = 0.5, CNVMean = 0.2)
}
\value{
Classification for LRR.
}
\description{
MakeOneMockSample: Run a single mock sample
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

