% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixOrList2df.R
\name{MatrixOrList2df}
\alias{MatrixOrList2df}
\title{MatrixOrList2df}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
MatrixOrList2df(tmp)
}
\arguments{
\item{tmp:}{Unknwon, default = Unknown.}
}
\value{
return data in a data frame.
}
\description{
MatrixOrList2df: Read data.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

