% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeCNVs.R
\name{MergeCNVs}
\alias{MergeCNVs}
\title{MergeCNVs}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
MergeCNVs(df, MaxNumSNPs = 50)
}
\arguments{
\item{Df:}{Data frame with predicted CNVs for each sample, default = Unknown.}

\item{MaxNumSNPs:}{Maximum number of SNPs per CNV, default = 50.}
}
\value{
Data frame with predicted CNVs.
}
\description{
MergeCNVs: Unknown.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

