% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MockData.R
\name{MockData}
\alias{MockData}
\title{MockData}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
MockData(N = 1, Wave1 = FALSE, Type = "Blood", Cores = 1)
}
\arguments{
\item{N:}{Unknown, default = 1.}

\item{Wave1:}{Unknown, default = FALSE.}

\item{Type:}{Unknown, default = Blood.}

\item{Cores:}{Number of cores used, default = 1.}
}
\value{
Data frame with predicted CNVs.
}
\description{
MockData: Unknown.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

