% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleMockData.R
\name{MultipleMockData}
\alias{MultipleMockData}
\title{MultipleMockData}
\usage{
MultipleMockData(NSamples = 10, NLoops = 10, Cores = 30,
  HMM = "/media/NeoScreen/NeSc_home/share/Programs/penncnv/lib/hhall.hmm",
  Path2PennCNV = "/media/NeoScreen/NeSc_home/share/Programs/penncnv/")
}
\value{
return data in data frame
}
\description{
MultipleMockData: Runs mock data multiple times.
}
\details{
Specifically designed to handle noisy data from amplified DNA on dried blood spot (DBS) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

