% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiplot.R
\name{Multiplot}
\alias{Multiplot}
\title{Multiplot}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
Multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{Plotlist:}{Unknown, default = NULL.}

\item{File:}{Unknown, default = Unknown.}

\item{Cols:}{Numeric, columns, default = 1.}

\item{Layout, }{Unknown, default = NULL.}
}
\value{
Unknown.
}
\description{
Multiplot: Unknown
}
\details{
Specifically designed to handle data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

