% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPlotsManySamples.R
\name{MultiPlotsManySamples}
\alias{MultiPlotsManySamples}
\title{MultiPlotsManySamples}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
MultiPlotsManySamples(df = iPsych_8_Samples_Good, Name = "Test.png",
  height = 1800, width = 2000, res = 120)
}
\arguments{
\item{df:}{Data frame with predicted CNVs for each sample, default = Unknown.}

\item{Name:}{Unknown, default =Unknown.}

\item{Height:}{Unknown, default = 1800.}

\item{Width:}{Unknown, default = 2000.}

\item{Res:}{Resolution, defualt = 120.}
}
\value{
Data frame with predicted CNVs.
}
\description{
MultiPlotsManySamples: Unknown.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

