% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NormalizeLRR.R
\name{NormalizeLRR}
\alias{NormalizeLRR}
\title{NormalizeLRR}
\usage{
NormalizeLRR(X, ExpectedMean = 0, DF = 40, NormQspline = TRUE)
}
\value{
LRR normalized
}
\description{
NormalizeLRR: Normalize LRR.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

