% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Plot.CNV.Info.R
\name{Plot.CNV.Info}
\alias{Plot.CNV.Info}
\title{Plot.CNV.Info}
\usage{
Plot.CNV.Info(MinNumSNPs, DF, subCNV, ID)
}
\value{
Data frame with Start and Stop.
}
\description{
Plot.CNV.Info: Define Start and Stop for CNV.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

