% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCNVs.R
\name{PlotCNVs}
\alias{PlotCNVs}
\title{PlotCNVs}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
PlotCNVs(DF, PathRawData = ".", Cores = 1, Skip = 10, PlotPosition = 1,
  Pattern = "", recursive = TRUE, dpi = 300, Files = NA, Start = NA,
  Stop = NA, SNPList = NULL, OutFolder = ".", Window = 35, key = NA)
}
\arguments{
\item{DF:}{Data frame with predicted CNVs for each sample, default = Unknown.}

\item{PathRawData:}{The path to the raw data files containing LRR and BAF values.}

\item{Cores:}{Number of cores used, default = 1.}

\item{Skip:}{Integer, the number of lines of the data file to be skipped before beginning to read the data, default = 0.}

\item{PlotPosition:}{Unknown, default = 1.}

\item{Pattern:}{File pattern in the raw data, default = "*".}

\item{Recursive:}{Logical, Unknown, default = TRUE.}

\item{Dpi:}{Dots per inch, default = 300.}

\item{Files:}{Unknown, default = NA.}

\item{SNPList:}{Getting chromosome (Chr) and position from another source than the RawFile - input should be the full path of the SNPList with columns: Name, Chr, and Position. Any positions from the RawFile will be erased. A PFB-column is also allowed but will be overwritten by the PFB-parameter or exchanged with 0.5, default = NULL.}

\item{Key:}{Exchange the ID printed on the plot and in the name of file with a deidentified ID - requires that the DF contains a column called ID_deidentified, default = NA.}

\item{Start:}{Start position of plot}

\item{Stop:}{Stop position of plot}

\item{OutFolder:}{Path for saving outputfiles, default is the current folder.}
}
\value{
One BAF- and LRR-plot for each CNV.
}
\description{
PlotCNVs: Function to plot Log R Ratio (LRR) and B Allele Frequency (BAF) of CNVs from a data frame (DF).
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
# Creating CNVs from MockData & plotting
MockCNVs <- MockData(N=2, Type="Blood", Cores=10)
CNVs <- iPsychCNV(PathRawData=".", Pattern="^MockSample*", Skip=0)
CNVs.Good <- subset(CNVs, CN != 2) # keep only CNVs with CN = 0, 1, 3, 4.
PlotCNVs(DF=CNVs.Good[1,], PathRawData=".", Cores=1, Skip=0, Pattern="^MockSamples*", key=NA, OutFolder="../", XAxisDefine = NULL)
}
\author{
Marcelo Bertalan, Ida Elken Sønderby, Louise K. Hoeffding.
}

