% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotHotspots.R
\name{PlotHotspots}
\alias{PlotHotspots}
\title{PlotHotspots}
\usage{
PlotHotspots(CNVsDF, Hotspots, AllHeadFiles, Cores = 1, Skip = 10,
  penalty = 60, Quantile = FALSE, QSpline = FALSE, sd = 0.18)
}
\value{
png files
}
\description{
PlotHotspots
}
\author{
Marcelo Bertalan
}

