% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots.R
\name{QC_Plots}
\alias{QC_Plots}
\title{QC_plots}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
QC_Plots(tmp)
}
\arguments{
\item{tmp:}{Unknown, default = Unknown.}
}
\value{
Unknown.
}
\description{
QC_plots: QC plots.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

