% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{QualityControl}
\alias{QualityControl}
\title{QualityControl}
\usage{
QualityControl(PathRawData = "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/",
  Cores = 10, Pattern = "*_*assed_*", NumFiles = "All", Skip = 0,
  Normalization = FALSE)
}
\value{
Data frame with QC variables.
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

