% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadCNV.R
\name{ReadCNV}
\alias{ReadCNV}
\title{ReadCNV}
\usage{
ReadCNV(RawFile = "Test.txt", skip = 0)
}
\value{
return data in data frame
}
\description{
ReadCNV: read data.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

