% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveCentromere.R
\name{RemoveCentromere}
\alias{RemoveCentromere}
\title{RemoveCentromere}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
RemoveCentromere(df, HG = "hg19")
}
\arguments{
\item{Df:}{Data frame, Unknown.}

\item{HG:}{Human genome version, default = hg19.}
}
\value{
Unknown.
}
\description{
RemoveCentromere: Remove centromeres from the analysis.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

