% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReScanCNVs.R
\name{ReScanCNVs}
\alias{ReScanCNVs}
\title{ReScanCNVs}
\usage{
ReScanCNVs(CNVs = CNVs,
  PathRawData = "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/", MINNumSNPs = 20,
  Cores = 1, hg = "hg18", NumFiles = "All", Pattern = "22q11_*",
  MinLength = 10, SelectedFiles = NA, Skip = 10, LCR = FALSE,
  PFB = NULL, chr = NA, penalty = 60, Quantile = FALSE,
  QSpline = FALSE, sd = 0.18, recursive = FALSE, CPTmethod = "meanvar",
  CNVSignal = 0.1, penvalue = 10, OutputPath = NA)
}
\value{
Data frame with CNVs predicted.
}
\description{
ReScanCNVs: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

