% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunGada.R
\name{RunGada}
\alias{RunGada}
\title{RunGada}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
RunGada(MockData = tmp, MinSegLen = 50, Tgada = 4.5)
}
\arguments{
\item{MockData:}{The mock data set, default = tmp.}

\item{MinSegLen:}{Unknown, default = 50.}

\item{Tgada:}{Unknown, default = 4.5.}
}
\value{
Data frame with predicted CNVs.
}
\description{
RunGada: Unknown.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

