% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSingleMock.R
\name{RunSingleMock}
\alias{RunSingleMock}
\title{RunSingleMock}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
RunSingleMock(Noise = 0, CNVMean = 0.4, Name = "Test",
  Method = "PennCNV")
}
\arguments{
\item{Noise:}{Unknown, default = 0.}

\item{CNVMean:}{Unknown, default = 0.4.}

\item{Name:}{Unknown, default =Unknown.}

\item{Method:}{The CNV predicting algorithm, default = PennCNV.}
}
\value{
Classification for Log R Ratio (LRR).
}
\description{
RunSingleMock: Run a single mock sample.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

