% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSamplesFromROI.R
\name{SelectSamplesFromROI}
\alias{SelectSamplesFromROI}
\title{SelectSamplesFromTOI}
\usage{
SelectSamplesFromROI(DF, roi, OverlapMin = 0.8, OverlapMax = 1.2)
}
\arguments{
\item{DF:}{Data frame with CNVs predicted for each sample.}

\item{roi:}{Regions of Interest or hotspots.}

\item{OverlapMin:}{Minimum overlap a CNV need to have with the hotspot to be selected.}

\item{OverlapMax:}{Maximum overlap accepted for a CNV and a hotspots.}
}
\value{
Data frame with CNVs predicted that belong to each hotspot.
}
\description{
SelectSamplesFromROI
}
\details{
Given hotspots regions and overlap cutoff it will return CNVs that belong to each hotspot.
}
\examples{
cnvs <- SelectSamplesFromROI(DF=Hotspots, roi=roi, OverlapMin=0.8, OverlapMax=1.2)
}
\author{
Marcelo Bertalan
}

