% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlidingWindowMean.R
\name{SlideWindowMean}
\alias{SlideWindowMean}
\title{SlideWindowMean}
\usage{
SlideWindowMean(X, N)
}
\value{
A vector with mean of the window size.
}
\description{
SlideWindowMean: Create a smooth effect on the data.
}
\details{
X <- a vector with the data. N <- the window size.
}
\author{
Marcelo Bertalan
}

