% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackPlot.R
\name{StackPlot}
\alias{StackPlot}
\title{Stacked plots of multiple sample of a specified loci}
\usage{
StackPlot(Pos, Ids, PathRawData, CNVs, Highlight = NULL)
}
\arguments{
\item{Pos}{Position of the loci to plot in the form chr21:1050000-1350000}

\item{PathRawData}{The path to the raw data files contining LRR and BAF values to plot}

\item{CNVs}{Data.frame containing the cnvs called on the samples}

\item{Highlight}{Position of a specific region to be highlighted, in the form chr21:1050000-1350000}

\item{IDs}{List of IDs to plot}
}
\value{
A png plot of the specified loci
}
\description{
stack.plot produces stacked plots of cnv calls from multiple samples
for a specific lucus. For each sample two tracks are plotted
representing the intensity and the B-allel frequency.
}
\examples{
mockCNV <- MockData(N=5, Type="Blood", Cores=1)
cnvs <- iPsychCNV(PathRawData=".", Cores=1, Pattern="^MockSample*", Skip=0)
StackPlot(Pos="chr21:28338230-46844965", Ids=unique(cnvs$ID), PathRawData=".", CNVs=cnvs, Highlight = "chr21:36653812-39117308")
}
\author{
Johan Hilge Thygesen
}

