% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BAF_drift.R
\name{BAF_drift}
\alias{BAF_drift}
\title{BAF_drift}
\usage{
BAF_drift(BAF)
}
\value{
Data frame with QC variables.
}
\description{
BAF drift: Calculate the drift in BAF.
}
\author{
Marcelo Bertalan
}

