% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAF_drift.R
\name{BAF_drift}
\alias{BAF_drift}
\title{BAF_drift}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
BAF_drift(BAF)
}
\arguments{
\item{BAF:}{BAF information from CNV region.}
}
\value{
Data frame with QC variables.
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays. 
Specifically designed to reduce false positive CNVs and handle data from amplified DNA on dried blood spots.
}
\details{
BAF drift: Calculate the drift in BAF. Used at PeaksAndPits function.
}
\examples{
BAF <- subCNV$B.Allele.Freq
BAF_Res <- BAF_drift(BAF)
}
\author{
Marcelo Bertalan
}

