% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckAbnChr.R
\name{CheckAbnChr}
\alias{CheckAbnChr}
\title{CheckAbnChr}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{

  CheckAbnChr(Path2RawFiles = "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/Version2",
  Cores = 1, Pattern = "*.txt$", skip = 10, NumFiles = "All")
}
\arguments{
\item{Path2RawFiles:}{Path for the Log R Ratio (LRR) and B Allele Frequency (BAF) files. Example: "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/Version2".}

\item{Cores:}{Number of cores to run in parallel, default = 1.}

\item{Pattern:}{Files pattern in the path. Example: "*.txt$".}

\item{Skip:}{Integer, the number of lines of the data file to be skipped before beginning to read data. Use if file has comments, default = 10.}

\item{NumFiles:}{Number of files to run. Example: numeric, 10 or character "All", default = All.}
}
\value{
Data frame with the estimate copy number for each chromosome.
}
\description{
CheckAbnChr: Estimate abnormal chromosome.
}
\details{
Specifically designed to reduce false positive CNVs and handle data from amplified DNA on dried blood spots.
}
\examples{
mockCNV <- MockData(N=5, Type="Blood", Cores=1)
cnvs <- CheckAbnChr(PathRawData=".", Cores=1, Pattern="^MockSample*", Skip=0)
}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

