% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckAbnChr.R
\name{CheckAbnChr}
\alias{CheckAbnChr}
\title{CheckAbnChr.}
\usage{

  CheckAbnChr(Path2RawFiles = "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/Version2",
  Cores = 40, Pattern = "*.txt$", skip = 10, NumFiles = "All")
}
\arguments{
\item{Path2RawFiles:}{Path where are the LLR and BAF files. Exaple: "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/Version2".}

\item{Cores:}{Number of cores to run in parallel.}

\item{Pattern:}{Files pattern in the path. Example: "*.txt$"}

\item{skip:}{How many rows should skip. Use if file has comments.}

\item{NumFiles:}{Number of files to run. Example: numeric, 10 or character "All".}
}
\value{
Data frame with the estimate copy number for each chromosome.
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays. 
Specifically designed to reduce false positive CNVs and handle data from amplified DNA on dried blood spots.
}
\details{
CheckAbnChr: Estimate abnormal chromosome.
}
\examples{
mockCNV <- MockData(N=5, Type="Blood", Cores=1)
cnvs <- CheckAbnChr(PathRawData=".", Cores=1, Pattern="^MockSample*", Skip=0)
}
\author{
Marcelo Bertalan
}

