% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassNumbers.R
\name{ClassNumbers}
\alias{ClassNumbers}
\title{ClassNumbers}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
ClassNumbers(tmpRaw)
}
\arguments{
\item{TmpRaw:}{Unknown.}
}
\value{
Data frame with a populate percentage for each class.
}
\description{
ClassNumbers: Define how B allele frequency (BAF) is distributed over five diferent classes.
}
\details{
Receive a vector with BAF and returns a data frame with a populate percentage for each class. Each class represent a group where a BAF is more likely to be found.
}
\examples{
Unknown.
}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

