% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CleaningPeaks.R
\name{CleaningPeaks}
\alias{CleaningPeaks}
\title{CleaningPeaks}
\usage{
CleaningPeaks(tp)
}
\value{
Peaks for BAF.
}
\description{
CleaningPeaks: Remove noise from peaks in BAF.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

