% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompressCNVs.R
\name{CompressCNVs}
\alias{CompressCNVs}
\title{CompressCNVs}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
CompressCNVs(tmp4, OverlapCutoff, Cores)
}
\arguments{
\item{tmp4:}{Unknown.}

\item{OverlapCutoff:}{Unknown.}

\item{Cores:}{Number of cores used, default = 1.}
}
\value{
Hotspots.
}
\description{
CompressCNVs: unknown
}
\details{
Unknown - description?
}
\examples{
Unknown
}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

