% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineBAFType.R
\name{DefineBAFType}
\alias{DefineBAFType}
\title{DefineBAFType}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
DefineBAFType(SumPeaks)
}
\arguments{
\item{SumPeaks:}{Unknown, default = Unknown.}
}
\value{
BAF classification.
}
\description{
DefineBAFType: Define B Allele Frequency (BAF) by peaks.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

