% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineBAFType.R
\name{DefineBAFType}
\alias{DefineBAFType}
\title{DefineBAFType}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
DefineBAFType(SumPeaks)
}
\value{
BAF classification.
}
\description{
DefineBAFType: Define BAF by peaks.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

