% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineCNVClass.R
\name{DefineCNVClass}
\alias{DefineCNVClass}
\title{DefineCNVClass}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
DefineCNVClass(tmp2)
}
\arguments{
\item{tmp2:}{Unknown}
}
\value{
CNV class.
}
\description{
DefineCNVClass: Define the CNV class.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Uknown

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

