% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DefineCNVClass.R
\name{DefineCNVClass}
\alias{DefineCNVClass}
\title{DefineCNVClass}
\usage{
DefineCNVClass(tmp2)
}
\value{
CNV class.
}
\description{
DefineCNVClass: Define if the CNV class.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

