% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineCNVType.R
\name{DefineCNVType}
\alias{DefineCNVType}
\title{DefineCNVType}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
DefineCNVType(tmp2)
}
\value{
CNV type.
}
\description{
DefineCNVType: Define if the CNV is good or bad.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

