% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineStartAndStop.R
\name{DefineStartAndStop}
\alias{DefineStartAndStop}
\title{DefineStartAndStop}
\usage{
DefineStartAndStop(indx, subCNV, CHR, ID, CPT.Res)
}
\value{
Data frame with Start and Stop.
}
\description{
DefineStartAndStop: Define Start and Stop for CNV.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

