% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefiningLogRRatio.R
\name{DefiningLogRRatio}
\alias{DefiningLogRRatio}
\title{DefiningLogRRatio}
\usage{
DefiningLogRRatio(res2)
}
\value{
Classification for LRR.
}
\description{
DefiningLogRRatio: Define LRR.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

