% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateMyBAF.R
\name{EvaluateMyBAF}
\alias{EvaluateMyBAF}
\title{EvaluateMyBAF}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
EvaluateMyBAF(res, res2)
}
\value{
Classification for BAF.
}
\description{
EvaluateMyBAF: Evaluate BAF.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

