% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCNVs.V4.R
\name{FilterCNVs.V4}
\alias{FilterCNVs.V4}
\title{FilterCNVs.V4.}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
FilterCNVs.V4(CNVs = CNVs, MinNumSNPs = 20, Sample, ID = "Test",
  verbose = FALSE)
}
\arguments{
\item{CNVs:}{Data frame with CNVs. Unknown?}

\item{MinNumSNPs:}{Minimum number of SNPs per CNV, default = 20.}

\item{Sample:}{Unknown.}

\item{ID:}{Unknown.}

\item{Verbose:}{Unknown, default = FALSE.}
}
\value{
Data frame with CNVs and classification.
}
\description{
FilterCNVs.V4.: Function to filter predicted Copy Number Variation (CNVs) and avoid a high number of false positive calls.
}
\details{
The function receives a data frame with CNV information, ex: chr., start position, stop position, and sample ID.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

