% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FilterCNVs.V4.R
\name{FilterCNVs.V4}
\alias{FilterCNVs.V4}
\title{iPsychCNVs}
\usage{
FilterCNVs.V4(CNVs = CNVs, MinNumSNPs, CNV, ID)
}
\value{
Data frame with CNVs and classification.
}
\description{
Filter CNVs: function to filter predicted CNVs and avoid high number of false positives.
}
\details{
The function receives a data frame with CNV information, ex: Chr, Start position, Stop position and Sample ID. Returns a data frame with CNVs and a Type difining if is bad or good.
}
\author{
Marcelo Bertalan
}

