% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindHighFreqCNVs.R
\name{FindHighFreqCNVs}
\alias{FindHighFreqCNVs}
\title{FindHighFreqCNVs}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
FindHighFreqCNVs(df, OverlapCutoff = 0.8)
}
\arguments{
\item{Df:}{Data frame with predicted CNVs for each sample.}

\item{OverlapCutoff:}{Unknown, default = 0.8.}
}
\value{
Data frame with predicted CNVs.
}
\description{
FindHighFreqCNVs: Identify CNVs with high frequency among the samples.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.
}
\author{
Marcelo Bertalan, Louise K. Hoeffidng.
}

