% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCNVClass.R
\name{GetCNVClass}
\alias{GetCNVClass}
\title{GetCNVClass}
\usage{
GetCNVClass(Variable)
}
\value{
return data in data frame
}
\description{
GetCNVClass: read data.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

