% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetLocus2.R
\name{GetLocus}
\alias{GetLocus}
\title{GetLocus}
\usage{
GetLocus(df)
}
\value{
data frame with locus
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

