% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLocus.R
\name{GetLocus}
\alias{GetLocus}
\title{GetLocus}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
GetLocus(df, hg = "hg19")
}
\arguments{
\item{df:}{Data frame (with predicted CNVs for each sample) with at least Chr, Start, Stop.}

\item{hg:}{Human genome version, default = hg19. For full range of possibilities, run ucscGenomes()$db {rtracklayer}}
}
\value{
data frame with CNV locus.
}
\description{
GetLocus: Adds locus to dataframe based on UCSC cytobands.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
GetLocus(CNVs, hg="hg18")

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

