% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HotspotsPrediction.R
\name{HotspotsPrediction}
\alias{HotspotsPrediction}
\title{HotspotsPrediction}
\usage{
HotspotsPrediction(hotspots, df)
}
\value{
return data in data frame
}
\description{
HotspotsPrediction: Count number of true CNVs in hotspots.
}
\author{
Marcelo Bertalan
}

