% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iPsychCNV.R
\name{iPsychCNV}
\alias{iPsychCNV}
\title{iPsychCNV}
\usage{
iPsychCNV(PathRawData = "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/",
  MINNumSNPs = 100, Cores = 10, NumFiles = "All", Pattern = "22q11_*",
  MinLength = 1e+05, SelectedFiles = NA, NormQspline = FALSE,
  DFspline = NA)
}
\value{
Data frame with CNVs predicted.
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

