% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPsychCNV.R
\name{iPsychCNV}
\alias{iPsychCNV}
\title{iPsychCNV}
\usage{
iPsychCNV(PathRawData = "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/",
  MINNumSNPs = 20, Cores = 1, hg = "hg19", NumFiles = "All",
  Pattern = "22q11_*", MinLength = 10, SelectedFiles = NA, Skip = 10,
  LCR = FALSE, PFB = NULL, chr = NA, penalty = 60, Quantile = FALSE,
  QSpline = FALSE, sd = 0.18, recursive = FALSE, CPTmethod = "meanvar",
  CNVSignal = 0.1, penvalue = 10, OutputPath = NA,
  OutputFileName = "Test", OnlyCNVs = FALSE)
}
\arguments{
\item{PathRawData}{The path to the raw data files contining LRR and BAF values.}
}
\value{
Data frame with CNVs predicted.
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
mockCNV <- MockData(N=5, Type="Blood", Cores=1)
cnvs <- iPsychCNV(PathRawData=".", Cores=1, Pattern="^MockSample*", Skip=0)
}
\author{
Marcelo Bertalan
}

