% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPsychCNV.R
\name{iPsychCNV}
\alias{iPsychCNV}
\title{iPsychCNV}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
iPsychCNV(PathRawData = "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/",
  MINNumSNPs = 20, Cores = 1, hg = "hg19", NumFiles = "All",
  Pattern = "22q11_*", MinLength = 10, SelectedFiles = NA, Skip = 10,
  LCR = FALSE, PFB = NULL, chr = NA, penalty = 60, Quantile = FALSE,
  QSpline = FALSE, sd = 0.18, recursive = FALSE, CPTmethod = "meanvar",
  CNVSignal = 0.1, penvalue = 16, OutputPath = NA,
  OutputFileName = "Test", OnlyCNVs = TRUE, SNPList = NULL)
}
\arguments{
\item{LCR}{list: Low copy repeat region, list of SNPs that should be removed.}

\item{PFB}{vector: Population frequency 0 to 1 for each SNP in the array.}

\item{PathRawData:}{The path to the raw data files contining Log R Ratio (LRR) and B Allele Frequency (BAF) values.}

\item{MINNumSNPs:}{Minimum number of SNPs per CNV, default = 20.}

\item{Cores:}{Number of cores used, default =  1.}

\item{Hg:}{Human genome version, default = hg19.}

\item{NumFiles:}{Number of files to be analyzed from PathRawData.}

\item{Pattern:}{File pattern in the PathRawData. Example: "*.txt".}

\item{MinLength:}{Minimum CNV length, default = Unknown.}

\item{SelectedFiles:}{List of file names that should be analyzed from PathRawData.}

\item{Skip:}{Integer, the number of lines of the data file to skip before beginning to read data.}

\item{Chr:}{Character, select a specific chromosome to be analyzed.}

\item{Penalty:}{The coefficient of the penalty for degrees of freedom in the GCV criterion. From smooth.spline {stats}.}

\item{Quantile:}{Logical, if quantile normalization should be applied or not, default = FALSE.}

\item{QSpline:}{Logical, if a cubic smoothing spline should be used to normalize the data, default = FALSE.}

\item{Sd:}{Numeric, LRR standard deviation for the quantile nomarlization, default = 0.18.}

\item{Recursive:}{Logical, should the listing recurse into directories? From list.files {base}.}

\item{CPTmethod:}{Character, method to find change points from changepoint package by Rebecca Killick. Default "meanvar", or "mean".}

\item{CNVSignal:}{Numeric, minumum CNV signal to be consider a CNV in absolute value, default = 0.1, any CNV with mean LRR in the CNV region with abs(X) < 0.1 is ignored.}

\item{Penvalue:}{Same as pen.value from function cpt.mean at changepoint R package by Rebecca Killick, default = 10. "The theoretical type I error e.g.0.05 when using the Asymptotic penalty.  A vector of length 2 (min,max) if using the CROPS penalty.  The value of the penalty when using the Manual penalty option - this can be a numeric value or text           giving the formula to use.  Available variables are, n=length of original data, null=null likelihood, alt=alternative likelihood, tau=proposed changepoint, diffparam=difference in number of alternatve and null parameters".}

\item{OutputPath:}{Character, path for output.}

\item{OutputFileName:}{Character, output file name.}

\item{OnlyCNVs:}{Logical, if TRUE only CNVs with copy number state 0,1,3,4 will be returned. If FALSE will return also changepoint regions with CN = 2.}

\item{SNPList:}{Getting Chr. and Position from another source than the RawFile - input should be the full path of the SNPList with columns: Name, Chr, amd Position. Any positions from the RawFile will be erased. A PFB-column is also allowed but will be overwritten by the PFB-parameter or exchanged with 0.5}
}
\value{
Data frame with predicted CNVs.
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
mockCNV <- MockData(N=5, Type="Blood", Cores=1)
cnvs <- iPsychCNV(PathRawData=".", Cores=1, Pattern="^MockSample*", Skip=0)
}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

