% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeOneMockSample.R
\name{MakeOneMockSample}
\alias{MakeOneMockSample}
\title{MakeOneMockSample}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
MakeOneMockSample(Noise = 0.5, CNVMean = 0.2)
}
\arguments{
\item{Noise:}{Unknown, defualt = 0.5.}

\item{CNVMean:}{Unknown, default = 0.2.}
}
\value{
Classification for Log R Ratio (LRR).
}
\description{
MakeOneMockSample: Run a single mock sample.
}
\details{
Specifically designed to handle noisy data from amplified DNA on phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
Unknown.

}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

