% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MatrixOrList2df.R
\name{MatrixOrList2df}
\alias{MatrixOrList2df}
\title{MatrixOrList2df}
\usage{
MatrixOrList2df(tmp)
}
\value{
return data in data frame
}
\description{
MatrixOrList2df: read data.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

