% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeCNVs.R
\name{MergeCNVs}
\alias{MergeCNVs}
\title{iPsychCNV}
\usage{
MergeCNVs(df, MaxNumSNPs = 50)
}
\value{
Data frame with CNVs predicted.
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

