% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MultiPlotsManySamples.R
\name{MultiPlotsManySamples}
\alias{MultiPlotsManySamples}
\title{iPsychCNV}
\usage{
MultiPlotsManySamples(df = iPsych_8_Samples_Good, Name = "Test.png",
  height = 1800, width = 2000, res = 120)
}
\value{
Data frame with CNVs predicted.
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

