% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAllCNVs.R
\name{PlotAllCNVs}
\alias{PlotAllCNVs}
\title{PlotAllCNVs}
\usage{
PlotAllCNVs(df = CNV.Res, Name = "CNV.Res.Test.png", NCOL = 2,
  Roi = roi, width = 16, height = 30, hg = "hg18")
}
\value{
return data in data frame
}
\description{
PlotAllCNVs: Plot all CNVs in one plot.
}
\author{
Marcelo Bertalan
}

