% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCNVsFromDataFrame.R
\name{PlotCNVsFromDataFrame}
\alias{PlotCNVsFromDataFrame}
\title{PlotCNVsFromDataFrame}
\usage{
PlotCNVsFromDataFrame(DF, PathRawData = ".", Cores = 1, Skip = 0,
  PlotPosition = 1, Pattern = "*", recursive = TRUE, dpi = 100,
  Files = NA)
}
\value{
Data frame with CNVs predicted.
}
\description{
Plot CNVs from data frame
}
\author{
Marcelo Bertalan
}

