% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotLRRAndCNVs.R
\name{PlotLRRAndCNVs}
\alias{PlotLRRAndCNVs}
\title{PlotLRRAndCNVs}
\usage{
PlotLRRAndCNVs(CNV, Sample = MockData, CNVMean = 0.3, Name = "Test.png",
  Roi = RoiSingleMock)
}
\value{
Classification for LRR.
}
\description{
PlotLRRAndCNVs: Plot  whole mock sample
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

