% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotXandYChr.R
\name{PlotXandYChr}
\alias{PlotXandYChr}
\title{PlotXandYChr}
\usage{
PlotXandYChr(Sample = MockData, Name = "Test.png")
}
\value{
png plot
}
\description{
PlotXandYChr
}
\author{
Marcelo Bertalan
}

