% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QC_plots.R
\name{QC_Plots}
\alias{QC_Plots}
\title{iPsychCNV}
\usage{
QC_Plots(tmp)
}
\value{
Data frame with CNVs predicted.
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

