% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadSample.R
\name{ReadSample}
\alias{ReadSample}
\title{ReadSample}
\usage{
ReadSample(RawFile = "Test.txt", skip = 0, LCR = FALSE, PFB = NULL,
  chr = NA)
}
\value{
return data in data frame
}
\description{
ReadSample: read data.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

