% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReScanCNVs.R
\name{ReScanCNVs}
\alias{ReScanCNVs}
\title{ReScanCNVs}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
ReScanCNVs(CNVs = CNVs,
  PathRawData = "/media/NeoScreen/NeSc_home/ILMN/iPSYCH/", MINNumSNPs = 5,
  Cores = 1, hg = "hg19", NumFiles = "All", Pattern = "*",
  MinLength = 10, SelectedFiles = NA, Skip = 10, LCR = FALSE,
  PFB = NULL, chr = NA, penalty = 60, Quantile = FALSE,
  QSpline = FALSE, sd = 0.18, recursive = FALSE, CPTmethod = "meanvar",
  CNVSignal = 0.1, penvalue = 10, OutputPath = NA, IndxPos = FALSE,
  ResPerSample = FALSE, Files = NA)
}
\arguments{
\item{CNVs}{data frame with hotspots to re-scan. Minimum information necessary is Chr, Start and Stop position.}

\item{PathRawData}{The path to the raw data files contining LRR and BAF values.}

\item{MINNumSNPs}{Minimum number of SNPs per CNV. Default 20.}

\item{Cores}{Number of cores used. Default 1.}

\item{hg}{Human genome version. Default hg19.}

\item{NumFiles}{Number of files to be analyzed from PathRawData.}

\item{Pattern}{File pattern in the PathRawData. Example: "*.txt".}

\item{MinLength}{Minimum CNV length.}

\item{SelectedFiles}{List of file names that should be analyzed from PathRawData.}

\item{Skip}{integer: the number of lines of the data file to skip before beginning to read data.}

\item{LCR}{list: low complex region, list of SNPs that should be removed.}

\item{PFB}{vector: Population frequency 0 to 1 for each SNP in the array.}

\item{chr}{character: select a specific chromosome to be analyzed.}

\item{penalty}{the coefficient of the penalty for degrees of freedom in the GCV criterion. From smooth.spline {stats}.}

\item{Quantile}{logical, if quantile normalization should be applied or not. Default FALSE.}

\item{QSpline}{logical, if a cubic smoothing spline should be used to normalize the data. Default FALSE.}

\item{sd}{numeric Log R ratio standard deviation for the quantile nomarlization. Default 0.18.}

\item{recursive}{logical, Should the listing recurse into directories? From list.files {base}.}

\item{CPTmethod}{character, method to find change points from changepoint package by Rebecca Killick. Default "meanvar", or "mean".}

\item{CNVSignal}{numeric, minumum CNV signal to be consider a CNV in absolute value. Default 0.1, any CNV with mean Log R ration in the CNV region with abs(X) < 0.1 is ignored.}

\item{penvalue}{Same as pen.value from function cpt.mean at changepoint R package by Rebecca Killick. Default 10. "The theoretical type I error e.g.0.05 when using the Asymptotic penalty.  A vector of length 2 (min,max) if using the CROPS penalty.  The value of the penalty when using the Manual penalty option - this can be a numeric value or text           giving the formula to use.  Available variables are, n=length of original data, null=null likelihood, alt=alternative likelihood, tau=proposed changepoint, diffparam=difference in number of alternatve and null parameters".}

\item{OutputPath}{character, path for output.}

\item{IndxPos, }{if index position for each hotspot is not included it will calculate. However, this step is time consuming.}

\item{ResPerSample, }{If TRUE saves the results from each sample.}

\item{Files, }{a vector with all samples name and path. If too many samples, list all files using recursive=TRUE can take long time.}

\item{OutputFileName}{character, Output file name.}

\item{OnlyCNVs}{logical, if TRUE only CNVs with copy number state 0,1,3,4 will be returned. If FALSE will return also changepoint regions with CN = 2.}
}
\value{
Data frame with CNVs predicted.
}
\description{
ReScanCNVs: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\examples{
MockDataCNVs <- MockData(N=100, Type="PKU", Cores=20)
iPsych.Pred <- iPsychCNV(PathRawData=".", MINNumSNPs=20, Cores=20, Pattern="^MockSample", MinLength=10, Skip=0)
iPsych.Pred.hotspots <- HotspotsCNV(iPsych.Pred, Freq=2, OverlapCutoff=0.9, Cores=1)
iPsych.Pred.Rescan <- ReScanCNVs(iPsych.Pred.hotspots, PathRawData="./Data", hg="hg19", Pattern="*", Cores=20, Skip=0)
iPsych.Pred.Rescan$ID <- iPsych.Pred.Rescan$SampleID
PlotAllCNVs(iPsych.Pred.Rescan, Name="iPsych.Pred.Rescan.png", hg="hg19", Roi=MockDataCNVs.roi)
}
\author{
Marcelo Bertalan
}

