% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RunLongMock.R
\name{RunLongMock}
\alias{RunLongMock}
\title{RunSingleMock}
\usage{
RunLongMock(Name = "Test", Method = "PennCNV", CNVDistance = 1000,
  Type = c(0, 1, 2, 3, 4), Mean = c(-0.3, -0.6, 0.3, 0.6), Size = c(300,
  600),
  HMM = "/media/NeoScreen/NeSc_home/share/Programs/penncnv/lib/hhall.hmm",
  Path2PennCNV = "/media/NeoScreen/NeSc_home/share/Programs/penncnv/")
}
\value{
Classification for LRR.
}
\description{
RunLongMock: Run a single mock sample
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

