% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RunPennCNV.R
\name{RunPennCNV}
\alias{RunPennCNV}
\title{iPsychCNV}
\usage{
RunPennCNV(PathRawData = "~/CNVs/MockData/PKU/Data", MINNumSNPs = 20,
  Pattern = ".*Mock.*\\\\.tab$", Cores = 20, Skip = 0,
  Normalization = FALSE, PFB = "NO",
  HMM = "/media/NeoScreen/NeSc_home/share/Programs/penncnv/lib/hhall.hmm",
  Path2PennCNV = "/media/NeoScreen/NeSc_home/share/Programs/penncnv/",
  penalty = 60, Quantile = TRUE, QSpline = TRUE, sd = 0.15)
}
\value{
Data frame with CNVs predicted.
}
\description{
iPsychCNV: Find Copy Number Variation (CNV) from SNP genotyping arrays.
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

