% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSingleMock.R
\name{RunSingleMock}
\alias{RunSingleMock}
\title{RunSingleMock}
\usage{
RunSingleMock(Noise = 0, CNVMean = 0.4, Name = "Test",
  Method = "PennCNV")
}
\value{
Classification for LRR.
}
\description{
RunSingleMock: Run a single mock sample
}
\details{
Specifically designed to handle noisy data from amplified DNA on  Phenylketonuria (PKU) cards. The function is a pipeline using many subfunctions.
}
\author{
Marcelo Bertalan
}

