% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSamplesFromROI.R
\name{SelectSamplesFromROI}
\alias{SelectSamplesFromROI}
\title{SelectSamplesFromROI}
\usage{
SelectSamplesFromROI(DF, roi, OverlapMin = 0.8, OverlapMax = 1.2,
  Cores = 1)
}
\arguments{
\item{DF:}{Data frame with predicted CNVs for each sample.}

\item{Roi:}{Regions Of Interest or hotspots.}

\item{OverlapMin:}{Minimum overlap a CNV need to have with the hotspot to be selected, default = 0.8.}

\item{OverlapMax:}{Maximum overlap accepted for a CNV and a hotspots, default = 1.2.}
}
\value{
Data frame with predicted CNVs that belong to each hotspot.
}
\description{
SelectSamplesFromROI: Select region of interest based on the samples.
}
\details{
Given hotspots regions and overlap cutoff it will return CNVs that belong to each hotspot.
}
\examples{
cnvs <- SelectSamplesFromROI(DF=Hotspots, roi=roi, OverlapMin=0.8, OverlapMax=1.2)
}
\author{
Marcelo Bertalan, Louise K. Hoeffding.
}

