% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackPlot.R
\name{StackPlot}
\alias{StackPlot}
\title{StackPlot}
\source{
\url{http://biopsych.dk/iPsychCNV}
}
\usage{
StackPlot(Pos, IDs, PathRawData, CNVs, Highlight = NULL, SNPList = NULL,
  key = NA, OutFolder = ".", Files = NA, Pattern = "",
  recursive = FALSE)
}
\arguments{
\item{Pos:}{Position of the loci to plot in the form chr21:1050000-1350000.}

\item{IDs:}{List of IDs to plot}

\item{PathRawData:}{The path to the raw data files contining LRR and BAF values to plot.}

\item{CNVs:}{Data frame containing the CNVs called on the samples.}

\item{Highlight:}{Position of a specific region to be highlighted, in the form chr21:1050000-1350000, default = NULL.}

\item{SNPList:}{Getting chromosome (chr) and position from another source than the RawFile - input should be the full path of the SNPList with columns: Name, Chr, and Position. Any positions from the RawFile will be erased. A PFB-column is also allowed but will be overwritten by the PFB-parameter or exchanged with 0.5, default = NULL.}

\item{key:}{Exchange the ID printed on the plot with a deidentified ID - requires that the CNV data frame contains a column called ID_deidentified, default = NA.}

\item{OutFolder:}{Path for saving outputfiles. Default is current folder}

\item{Files:}{Full path for each sample. You can get it using the command: Files <- list.files(path=PathRawData, pattern=Pattern, full.names=TRUE, recursive=recursive). Then use Files=Files.}

\item{Pattern:}{File pattern in the raw data, default = "*".}

\item{Recursive:}{Logical, Should files matching patterns in subfolders be included, default = FALSE.}
}
\value{
A png plot of the specified loci.
}
\description{
Stackplot: Produces stacked plots of CNV calls from multiple samples for a given locus.
}
\details{
For each sample two tracks are plotted representing the Log R Ratio (LRR) and the B Allele Frequency (BAF).
}
\examples{
mockCNV <- MockData(N=5, Type="Blood", Cores=1)
cnvs <- iPsychCNV(PathRawData=".", Cores=1, Pattern="^MockSample*", Skip=0)
StackPlot(Pos="chr21:28338230-46844965", IDs=unique(cnvs$ID), PathRawData=".", CNVs=cnvs, Highlight = "chr21:36653812-39117308")
}
\author{
Johan Hilge Thygesen, Ida Elken Sønderby, Louise K. Hoeffding.
}

